added initial blowdart functionality
