+++ Assets/Animations/PlayerAnimations/Default/Player.controller
@@ -124,7 +124,7 @@ AnimatorStateTransition:
   m_Name: 
   m_Conditions:
   - m_ConditionMode: 1
-    m_ConditionEvent: flamethrowerHeld
+    m_ConditionEvent: weaponHeld
     m_EventTreshold: 0
   - m_ConditionMode: 2
     m_ConditionEvent: mouseBehind
@@ -265,7 +265,7 @@ AnimatorStateTransition:
   m_Name: 
   m_Conditions:
   - m_ConditionMode: 1
-    m_ConditionEvent: flamethrowerHeld
+    m_ConditionEvent: weaponHeld
     m_EventTreshold: 0
   - m_ConditionMode: 1
     m_ConditionEvent: mouseBehind
@@ -389,7 +389,7 @@ AnimatorStateTransition:
   m_Name: 
   m_Conditions:
   - m_ConditionMode: 2
-    m_ConditionEvent: flamethrowerHeld
+    m_ConditionEvent: weaponHeld
     m_EventTreshold: 0
   m_DstStateMachine: {fileID: 0}
   m_DstState: {fileID: -8359475252742692441}
@@ -551,7 +551,7 @@ AnimatorController:
     m_DefaultInt: 0
     m_DefaultBool: 0
     m_Controller: {fileID: 9100000}
-  - m_Name: flamethrowerHeld
+  - m_Name: weaponHeld
     m_Type: 4
     m_DefaultFloat: 0
     m_DefaultInt: 0
@@ -613,7 +613,7 @@ AnimatorStateTransition:
   m_Name: 
   m_Conditions:
   - m_ConditionMode: 2
-    m_ConditionEvent: flamethrowerHeld
+    m_ConditionEvent: weaponHeld
     m_EventTreshold: 0
   m_DstStateMachine: {fileID: 0}
   m_DstState: {fileID: -8359475252742692441}
