--- Assets/Scripts/Player/PlayerWeaponSwitcher.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 89856b9df22be4b3ab09bfd4bfdceafe
