--- Assets/Scripts/Weapons/FlameThrower/Weapon.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9f0a389d2f38045d18a968d3b229f301
