--- Assets/Scripts/Weapons/FlameThrower/WeaponManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 59fbf05a44640451cabc587bc2e9a3bb
