+++ Assets/Scripts/Weapons/PlayerWeaponSwitcher.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 106c2c0958036f841a87d687f1f638a8
