+++ Assets/Scripts/Weapons/Weapon.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4e54d0da392a66843b2064dd2763b054
