+++ Assets/Scripts/Weapons/WeaponManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 1444a461b04eccf45b32db7b7a389a90
